dicts crud filter with user

+++ src/services/dict.rs
@@ -5,10 +5,10 @@ use crate::{
     entities::{dicts, prelude::Dicts},
 };
 use chrono::Utc;
-use sea_orm::{ActiveModelTrait, EntityTrait, Set};
+use sea_orm::{ActiveModelTrait, ColumnTrait, EntityTrait, QueryFilter, Set};
 use uuid::Uuid;
 
-pub async fn add_dict(req: DictAddRequest) -> AppResult<DictResponse> {
+pub async fn add_dict(req: DictAddRequest, user_id: String) -> AppResult<DictResponse> {
     let db = INTERNAL_DICT_DB
         .get()
         .ok_or(anyhow::anyhow!("数据库连接失败。"))?;
@@ -18,6 +18,8 @@ pub async fn add_dict(req: DictAddRequest) -> AppResult<DictResponse> {
         name: Set(req.name.clone()),
         language: Set(req.language.clone()),
         word_count: Set(0),
+        builtin: Set(false),
+        user_id: Set(user_id),
         created_at: Set(Utc::now().naive_utc()),
         updated_at: Set(Utc::now().naive_utc()),
     };
@@ -30,7 +32,7 @@ pub async fn add_dict(req: DictAddRequest) -> AppResult<DictResponse> {
     })
 }
 
-pub async fn update_dict(req: DictUpdateRequest) -> AppResult<DictResponse> {
+pub async fn update_dict(req: DictUpdateRequest, user_id: String) -> AppResult<DictResponse> {
     let db = INTERNAL_DICT_DB
         .get()
         .ok_or(anyhow::anyhow!("数据库连接失败。"))?;
@@ -39,11 +41,20 @@ pub async fn update_dict(req: DictUpdateRequest) -> AppResult<DictResponse> {
     if dict.is_none() {
         return Err(anyhow::anyhow!("词典不存在。").into());
     }
+    if dict.as_ref().unwrap().user_id != user_id {
+        return Err(anyhow::anyhow!("无权操作。").into());
+    }
     let mut dict: dicts::ActiveModel = dict.unwrap().into();
 
-    dict.name = Set(req.name.to_owned());
-    dict.language = Set(req.language.to_owned());
-    dict.word_count = Set(req.word_count);
+    if let Some(name) = req.name {
+        dict.name = Set(name);
+    }
+    if let Some(language) = req.language {
+        dict.language = Set(language);
+    }
+    if let Some(word_count) = req.word_count {
+        dict.word_count = Set(word_count);
+    }
     dict.updated_at = Set(Utc::now().naive_utc());
 
     let dict: dicts::Model = dict.update(db).await?;
@@ -56,19 +67,32 @@ pub async fn update_dict(req: DictUpdateRequest) -> AppResult<DictResponse> {
     })
 }
 
-pub async fn delete_dict(id: String) -> AppResult<()> {
+pub async fn delete_dict(id: String, user_id: String) -> AppResult<()> {
     let db = INTERNAL_DICT_DB
         .get()
         .ok_or(anyhow::anyhow!("数据库连接失败。"))?;
+
+    let dict = Dicts::find_by_id(id.clone())
+        .one(db)
+        .await?
+        .ok_or(anyhow::anyhow!("字典未找到。"))?;
+
+    // 校验 user_id
+    if dict.user_id != user_id {
+        return Err(anyhow::anyhow!("无权限删除该字典。").into());
+    }
     Dicts::delete_by_id(id).exec(db).await?;
     Ok(())
 }
 
-pub async fn dicts() -> AppResult<Vec<DictResponse>> {
+pub async fn dicts(user_id: String) -> AppResult<Vec<DictResponse>> {
     let db = INTERNAL_DICT_DB
         .get()
         .ok_or(anyhow::anyhow!("数据库连接失败。"))?;
-    let dicts = Dicts::find().all(db).await?;
+    let dicts = Dicts::find()
+        .filter(dicts::Column::UserId.eq(user_id))
+        .all(db)
+        .await?;
     let res = dicts
         .into_iter()
         .map(|dict| DictResponse {
